Move FOOOFGroup out to its own file
